+++ packages/sdk/src/transactionManager/transactionManager.ts
@@ -1,4 +1,4 @@
-import { BigNumber, constants, providers } from "ethers";
+import { BigNumber, constants, providers, Contract } from "ethers";
 import {
   PrepareParams,
   CancelParams,
@@ -8,12 +8,14 @@ import {
   createLoggingContext,
 } from "@connext/nxtp-utils";
 import { TransactionManager as TTransactionManager, IERC20Minimal } from "@connext/nxtp-contracts/typechain";
+import { parseError } from "@connext/nxtp-txservice";
 import TransactionManagerArtifact from "@connext/nxtp-contracts/artifacts/contracts/TransactionManager.sol/TransactionManager.json";
 import ERC20 from "@connext/nxtp-contracts/artifacts/contracts/interfaces/IERC20Minimal.sol/IERC20Minimal.json";
 import { Interface } from "ethers/lib/utils";
 import contractDeployments from "@connext/nxtp-contracts/deployments.json";
 
-import { ChainNotConfigured } from "../error";
+import { ChainNotConfigured, PriceOracleNotConfigured } from "../error";
+import { getDecimals, getTokenPrice } from "../utils";
 
 /**
  * Returns the address of the `TransactionManager` deployed to the provided chain, or undefined if it has not been deployed
@@ -32,6 +34,42 @@ export const getDeployedTransactionManagerContract = (chainId: number): { addres
 };
 
 /**
+ * Returns the address of the `PriceOracle` deployed to the provided chain, or undefined if it has not been deployed
+ *
+ * @param chainId - The chain you want the address on
+ * @returns The deployed address or `undefined` if it has not been deployed yet
+ */
+export const getDeployedPriceOracleContract = (chainId: number): { address: string; abi: any } | undefined => {
+  const record = (contractDeployments as any)[String(chainId)] ?? {};
+  const name = Object.keys(record)[0];
+  if (!name) {
+    return undefined;
+  }
+  const contract = record[name]?.contracts?.ConnextPriceOracle;
+  return { address: contract.address, abi: contract.abi };
+};
+
+/**
+ * Returns the addresses where the price oracle contract is deployed to
+ *
+ */
+export const getDeployedChainIdsForGasFee = (): number[] => {
+  const chainIdsForGasFee: number[] = [];
+  const chainIds = Object.keys(contractDeployments);
+  chainIds.forEach((chainId) => {
+    const record = (contractDeployments as any)[String(chainId)];
+    const chainName = Object.keys(record)[0];
+    if (chainName) {
+      const priceOracleContract = record[chainName]?.contracts?.ConnextPriceOracle;
+      if (priceOracleContract) {
+        chainIdsForGasFee.push(Number(chainId));
+      }
+    }
+  });
+  return chainIdsForGasFee;
+};
+
+/**
  * @classdesc Multi-chain wrapper around TranasctionManager contract interactions
  */
 export class TransactionManager {
@@ -39,6 +77,7 @@ export class TransactionManager {
     [chainId: number]: {
       provider: providers.FallbackProvider;
       transactionManagerAddress: string;
+      priceOracleAddress: string;
     };
   };
 
@@ -50,16 +89,18 @@ export class TransactionManager {
       [chainId: number]: {
         provider: providers.FallbackProvider;
         transactionManagerAddress: string;
+        priceOracleAddress: string;
       };
     },
     private readonly signerAddress: Promise<string>,
     private readonly logger: Logger,
   ) {
     this.chainConfig = {};
-    Object.entries(_chainConfig).forEach(([chainId, { provider, transactionManagerAddress }]) => {
+    Object.entries(_chainConfig).forEach(([chainId, { provider, transactionManagerAddress, priceOracleAddress }]) => {
       this.chainConfig[parseInt(chainId)] = {
         transactionManagerAddress,
         provider,
+        priceOracleAddress,
       };
     });
   }
@@ -330,5 +371,110 @@ export class TransactionManager {
     const [balance] = this.txManagerInterface.decodeFunctionResult("routerBalances", encoded);
 
     return BigNumber.from(balance);
+  }
+
+  /**
+   * Calculates gas amount in receiving token
+   *
+   * @param chainId - The receiving chain you want to check gas price on
+   * @param fulfillParams - The parameters for fulfill transactions
+   */
+  async calculateGasInTokenForFullfil(
+    chainId: number,
+    fulfillParams: FulfillParams,
+    _requestContext?: RequestContext<string>,
+  ): Promise<BigNumber> {
+    const { requestContext, methodContext } = createLoggingContext(
+      "TransactionManager.calculateGasInToken",
+      _requestContext,
+      fulfillParams.txData.transactionId,
+    );
+
+    const { transactionManagerAddress, provider } = this.chainConfig[chainId] ?? {};
+    if (!transactionManagerAddress || !provider) {
+      throw new ChainNotConfigured(chainId, Object.keys(this.chainConfig));
+    }
+
+    this.logger.info("Method start", requestContext, methodContext, { fulfillParams });
+    const gasAmount = await this.calculateGasAmountForFulfill(chainId, fulfillParams);
+    const { txData } = fulfillParams;
+
+    const ethPriceInUsd = await getTokenPrice(
+      this.chainConfig[chainId]?.priceOracleAddress,
+      constants.AddressZero,
+      provider,
+    );
+
+    if (ethPriceInUsd.isZero()) {
+      throw new PriceOracleNotConfigured(chainId, constants.AddressZero);
+    }
+
+    const receivingTokenPriceInUsd = await getTokenPrice(
+      this.chainConfig[chainId]?.priceOracleAddress,
+      txData.receivingAssetId,
+      provider,
+    );
+
+    if (receivingTokenPriceInUsd.isZero()) {
+      throw new PriceOracleNotConfigured(chainId, txData.receivingAssetId);
+    }
+
+    const outputDecimals = await getDecimals(txData.receivingAssetId, provider);
+
+    const tokenAmount = gasAmount
+      .mul(ethPriceInUsd)
+      .div(receivingTokenPriceInUsd)
+      .div(BigNumber.from(10).pow(18 - outputDecimals));
+
+    return tokenAmount;
+  }
+
+  /**
+   * Calculates gas amount for fulfill tranactions
+   *
+   * @param chainId The network identifier
+   * @param fulfillParams The params used for fulfill transactions
+   */
+  async calculateGasAmountForFulfill(chainId: number, fulfillParams: FulfillParams): Promise<BigNumber> {
+    const { transactionManagerAddress, provider } = this.chainConfig[chainId] ?? {};
+    if (!transactionManagerAddress || !provider) {
+      throw new ChainNotConfigured(chainId, Object.keys(this.chainConfig));
+    }
+
+    const transactionManager = new Contract(
+      transactionManagerAddress,
+      TransactionManagerArtifact.abi,
+      provider,
+    ) as TTransactionManager;
+
+    const { txData, relayerFee, signature, callData } = fulfillParams;
+
+    const contractArgs = {
+      txData,
+      relayerFee,
+      signature,
+      callData,
+      encodedMeta: "0x",
+    };
+
+    // get gas limit
+    let gasLimit = BigNumber.from(0);
+    try {
+      gasLimit = await transactionManager.estimateGas.fulfill(contractArgs);
+    } catch (e) {
+      const sanitized = parseError(e);
+      throw sanitized;
+    }
+
+    // get gas price
+    let gasPrice = BigNumber.from(0);
+    try {
+      gasPrice = await provider.getGasPrice();
+    } catch (e) {
+      const sanitized = parseError(e);
+      throw sanitized;
+    }
+
+    return gasLimit.mul(gasPrice);
   }
 }