+++ CHANGELOG.md
@@ -2,9 +2,18 @@
 
 ## Next Release
 
+## 0.0.58
+
+- [subgraphs] Fix user typo
+- [txservice] Sync chain providers
+- [sdk] Estimate gas on fulfill for configured chains
+
+## 0.0.57
+
 - [txservice] Add nonce gap backfill in event that txcount backtracks
 - [txservice] Curb gas spikes
 - [sdk/router] Catch errors per chain in active tx subgraph poller
+- [router] Fix estimate gas bug
 
 ## 0.0.56